+++ app/src/terminal/local_tty/mod.rs
@@ -29,9 +29,9 @@ use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use shell::ShellStarter;
 
-#[cfg(windows)]
-pub use self::terminal_manager::shutdown_all_pty_event_loops;
 pub use self::terminal_manager::{get_shell_starter, TerminalManager};
+#[cfg(windows)]
+pub use self::terminal_view_adaptor::shutdown_all_pty_event_loops;
 pub(crate) use self::terminal_view_adaptor::{
     create_terminal_view_surface, terminal_view_restored_blocks, TerminalViewSurfaceConfig,
 };
